servers/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
python-dotenv>=1.0.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
pymysql>=1.1.0,<2.0.0
httpx>=0.27.0,<0.28.0
opencv-python>=4.9.0,<5.0.0
